refactor loops
