--- src/user_loop.rs
@@ -1,2095 +1,0 @@
-use super::input;
-use super::rv_api;
-use super::utils;
-use crate::rv_api::get_box_info_admin;
-use crate::rv_api::get_product_info;
-use crate::rv_api::get_user_info;
-use crate::rv_api::return_product;
-use crate::rv_api::update_box;
-use crate::rv_api::ApiResultPurchaseItem;
-use crate::rv_api::ApiResultPurchaseItemFailType;
-use crate::rv_api::ApiResultValue;
-use crate::rv_api::ProductCategory;
-use crate::rv_api::UserInfoTrait;
-use crate::utils::clear_terminal;
-use crate::utils::load_ascii;
-use crate::utils::print_error_line;
-use crate::utils::print_title;
-use crate::utils::printline;
-use crate::utils::purchase_fail_bell;
-use crate::utils::readline;
-use crate::utils::TimeoutResult;
-use crate::TerminalIO;
-use crate::INPUT_TIMEOUT_LONG;
-use crate::INPUT_TIMEOUT_SHORT;
-
-use chrono::{DateTime, Local};
-use crossterm::{
-    cursor,
-    event::{Event, KeyCode},
-    execute, queue,
-    style::{Color, Print, PrintStyledContent, Stylize},
-    terminal::{self, disable_raw_mode},
-    ExecutableCommand,
-};
-use input::InputEvent;
-use regex::Regex;
-use rv_api::ApiResult;
-use std::process::exit;
-use std::sync::mpsc::RecvTimeoutError;
-use std::sync::LazyLock;
-use std::thread::sleep;
-use std::time::Duration;
-static PURCHASE_FAILED_MSG1: LazyLock<String> = load_ascii!("../ascii/purchase_failed.txt");
-static PURCHASE_FAILED_MSG2: LazyLock<String> = load_ascii!("../ascii/purchase_failed2.txt");
-static COFFEE_MSG: LazyLock<String> = load_ascii!("../ascii/netlight.txt");
-
-fn new_product(
-    barcode: &str,
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    let name;
-    utils::printline(
-        terminal_io,
-        &format!("Creating a new product. Enter to cancel."),
-    );
-    utils::printline(terminal_io, &format!("Enter product name:"));
-
-    let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-        TimeoutResult::RESULT(s) => s,
-    };
-    if input_line.len() > 0 {
-        name = input_line.to_string();
-    } else {
-        printline(terminal_io, "Cancelled.");
-        return TimeoutResult::RESULT(());
-    }
-
-    printline(terminal_io, "");
-
-    let buy_price = loop {
-        utils::printline(
-            terminal_io,
-            "Enter item buyprice. Format: [0-9]+\\.[0-9][0-9]",
-        );
-        utils::printline(terminal_io, "At least one number, followed by period, followed by two numbers. For example: '1.00', '0.01', '14.42'");
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            printline(terminal_io, "Cancelled.");
-            return TimeoutResult::RESULT(());
-        } else if Regex::new("^[0-9]+\\.[0-9][0-9]$")
-            .expect("")
-            .is_match(&input_line)
-        {
-            break input_line.replace(".", "").parse().unwrap();
-        } else {
-            print_error_line(terminal_io, "Invalid price entered, please retry!\n");
-        }
-    };
-    printline(terminal_io, "");
-
-    let sell_price = loop {
-        utils::printline(
-            terminal_io,
-            "\r\nEnter item sellprice. Format: [0-9]+\\.[0-9][0-9]",
-        );
-        utils::printline(terminal_io, "At least one number, followed by period, followed by two numbers. For example: '1.00', '0.01', '14.42'");
-        let margin = rv_api::get_margin(&credentials).unwrap() as f64;
-        let margin_pretty = format!("{}%", (margin * 100.0).ceil());
-        let suggested_price = (buy_price as f64 * (1.0 + margin)).ceil() as i32;
-        utils::printline(
-            terminal_io,
-            &format!(
-                "Suggest {} calculated with the margin of {}",
-                &utils::format_money(&suggested_price),
-                margin_pretty
-            ),
-        );
-        utils::printline(
-            terminal_io,
-            &format!(
-                "Modify or keep [{}]:",
-                &utils::format_money(&suggested_price)
-            ),
-        );
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            printline(terminal_io, "Using the suggested price.");
-            break suggested_price;
-        } else if Regex::new("^[0-9]+\\.[0-9][0-9]$")
-            .expect("")
-            .is_match(&input_line)
-        {
-            break input_line.replace(".", "").parse().unwrap();
-        } else {
-            print_error_line(terminal_io, "Invalid price entered, please retry!\n");
-        }
-    };
-    printline(terminal_io, "");
-    let stock = loop {
-        let suggested_stock = 0;
-        utils::printline(
-            terminal_io,
-            "Enter item stock. Format: [0-9]+ or [0.9]+\\*[0.9]+",
-        );
-        utils::printline(terminal_io, &format!("Modify or keep [{suggested_stock}]"));
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            printline(terminal_io, "Nothing changed.");
-            break suggested_stock;
-        }
-        match utils::calculator_input(&input_line) {
-            Some(stock) => {
-                break stock;
-            }
-            None => {
-                print_error_line(terminal_io, "Invalid stock entered, please retry!\n");
-            }
-        }
-    };
-    printline(terminal_io, "");
-
-    let category = loop {
-        utils::printline(terminal_io, "Enter product category id.");
-        utils::printline(terminal_io, "Categories available:");
-        let categories = rv_api::get_categories(&credentials).unwrap();
-        for category in categories.iter() {
-            utils::printline(
-                terminal_io,
-                &format!("{}, id: {}", category.description, category.category_id),
-            );
-        }
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            print_error_line(terminal_io, "Invalid category id entered, please retry!\n");
-            continue;
-        } else if Regex::new("^[0-9]+").expect("").is_match(&input_line) {
-            let chosen: i32 = input_line.parse().unwrap();
-            match categories.iter().find(|c| c.category_id == chosen) {
-                Some(c) => {
-                    break ProductCategory {
-                        description: c.description.clone(),
-                        category_id: c.category_id,
-                    };
-                }
-                None => {
-                    print_error_line(terminal_io, "Invalid category id entered, please retry!\n");
-                    continue;
-                }
-            }
-        } else {
-            print_error_line(terminal_io, "Invalid category entered, please retry!\n");
-        }
-    };
-    match rv_api::add_product(
-        barcode,
-        &name,
-        category.category_id,
-        buy_price,
-        sell_price,
-        stock,
-        credentials,
-    )
-    .unwrap()
-    {
-        ApiResult::Success => utils::printline(terminal_io, "Product added."),
-        ApiResult::Fail(msg) => utils::print_error_line(terminal_io, &msg),
-    }
-    TimeoutResult::RESULT(())
-}
-
-fn new_box(
-    barcode: &str,
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    utils::printline(terminal_io, "Creating a new box.");
-    let product_barcode = loop {
-        utils::printline(terminal_io, "Enter product barcode.");
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            utils::printline(terminal_io, "Cancelled.");
-            return TimeoutResult::RESULT(());
-        } else if Regex::new("^[0-9]+$").expect("").is_match(&input_line) {
-            let product_barcode = input_line;
-            match rv_api::get_product_info(credentials, &product_barcode) {
-                Some(product_info) => {
-                    utils::printline(
-                        terminal_io,
-                        &format!(
-                            "Found an existing product with the given barcode: {}",
-                            product_info.name
-                        ),
-                    );
-                }
-                None => {
-                    utils::printline(
-                        terminal_io,
-                        "Couldn't find an existing product with the given barcode.",
-                    );
-                    utils::printline(terminal_io, "");
-                    if let TimeoutResult::TIMEOUT =
-                        new_product(&product_barcode, terminal_io, credentials)
-                    {
-                        return TimeoutResult::TIMEOUT;
-                    }
-                    if get_product_info(credentials, &product_barcode).is_none() {
-                        print_error_line(terminal_io, "Adding new product failed!");
-                        return TimeoutResult::RESULT(());
-                    }
-                }
-            }
-            break product_barcode;
-        } else {
-            print_error_line(terminal_io, "Invalid barcode entered, please retry!\n");
-        }
-    };
-    printline(terminal_io, "");
-
-    let items_per_box;
-    loop {
-        utils::printline(
-            terminal_io,
-            "Enter number of products in a box. Format: [0-9]+",
-        );
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if Regex::new("^[0-9]+").expect("").is_match(&input_line) {
-            items_per_box = input_line.parse().unwrap();
-            break;
-        } else {
-            print_error_line(terminal_io, "Invalid number entered, please retry!\n");
-        }
-    }
-    printline(terminal_io, "");
-
-    match rv_api::add_box(barcode, &product_barcode, items_per_box, credentials).unwrap() {
-        ApiResult::Success => {
-            utils::printline(terminal_io, &format!("Box added."));
-            utils::printline(terminal_io, &format!(""));
-            return buy_in_box(barcode, terminal_io, credentials);
-        }
-        ApiResult::Fail(msg) => print_error_line(terminal_io, &msg),
-    }
-    TimeoutResult::RESULT(())
-}
-
-fn new_item(
-    barcode: &str,
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    printline(
-        terminal_io,
-        &format!("Add a new box or product? [bp] or Enter to cancel."),
-    );
-    loop {
-        match terminal_io.recv.recv_timeout(INPUT_TIMEOUT_LONG) {
-            Err(RecvTimeoutError::Timeout) => return TimeoutResult::TIMEOUT,
-            Ok(input::InputEvent::Terminal(Event::Key(ev))) => match ev.code {
-                KeyCode::Enter => {
-                    utils::printline(terminal_io, "Cancelled!");
-                    return TimeoutResult::RESULT(());
-                }
-                KeyCode::Char(c) => match c.to_ascii_lowercase() {
-                    'b' => {
-                        utils::printline(terminal_io, "");
-                        return new_box(barcode, terminal_io, credentials);
-                    }
-                    'p' => {
-                        utils::printline(terminal_io, "");
-                        return new_product(barcode, terminal_io, credentials);
-                    }
-                    _ => (),
-                },
-                _ => (),
-            },
-            _ => (),
-        }
-    }
-}
-
-fn buy_in_box(
-    barcode: &str,
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    let box_ = match rv_api::get_box_info_admin(barcode, credentials).unwrap() {
-        Some(b) => b,
-        None => {
-            print_error_line(
-                terminal_io,
-                &format!("Buy in error: No box found with barcode {}", barcode),
-            );
-            return TimeoutResult::RESULT(());
-        }
-    };
-    utils::printline(
-        terminal_io,
-        &format!(
-            "Found a box containing {}x of {}",
-            box_.items_per_box, box_.product.name
-        ),
-    );
-
-    utils::printline(terminal_io, "Adding new box to stock.");
-    let mut buy_price = box_.product.buy_price;
-    let mut buy_price_changed = false;
-
-    utils::printline(
-        terminal_io,
-        &format!(
-            "Current item buyprice: {} x {} = {}",
-            &utils::format_money(&buy_price),
-            box_.items_per_box,
-            &utils::format_money(&(box_.items_per_box * buy_price))
-        ),
-    );
-    loop {
-        utils::printline(
-            terminal_io,
-            "Enter box buyprice. Format: [0-9]+\\.[0-9][0-9]",
-        );
-        utils::printline(terminal_io, "At least one number, followed by period, followed by two numbers. For example: '1.00', '0.01', '14.42'");
-
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            printline(terminal_io, "Nothing changed.");
-            break;
-        } else if Regex::new("^[0-9]+\\.[0-9][0-9]$")
-            .expect("")
-            .is_match(&input_line)
-        {
-            let box_buy_price = input_line.replace(".", "").parse::<i32>().unwrap();
-            buy_price = (box_buy_price as f64 / box_.items_per_box as f64).ceil() as i32;
-            buy_price_changed = true;
-
-            utils::printline(
-                terminal_io,
-                &format!(
-                    "Calculated new item buyprice: {}",
-                    &utils::format_money(&buy_price)
-                ),
-            );
-            break;
-        } else {
-            print_error_line(terminal_io, "Invalid price entered, please retry!\n");
-        }
-    }
-    printline(terminal_io, "");
-
-    let mut sell_price = box_.product.sell_price;
-    loop {
-        utils::printline(terminal_io, "\r\nEnter item sellprice.");
-        if buy_price_changed {
-            let margin = rv_api::get_margin(&credentials).unwrap() as f64;
-            let margin_pretty = format!("{}%", (margin * 100.0).ceil());
-            sell_price = (buy_price as f64 * (1.0 + margin)).ceil() as i32;
-            utils::printline(
-                terminal_io,
-                &format!(
-                    "Suggest {} calculated with the margin of {}",
-                    &utils::format_money(&sell_price),
-                    margin_pretty
-                ),
-            );
-        }
-        utils::printline(
-            terminal_io,
-            &format!("Modify or keep [{}]:", &utils::format_money(&sell_price)),
-        );
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            if buy_price_changed {
-                printline(terminal_io, "Using the suggested price.");
-            } else {
-                printline(terminal_io, "Nothing changed.");
-            }
-            break;
-        } else if Regex::new("^[0-9]+\\.[0-9][0-9]$")
-            .expect("")
-            .is_match(&input_line)
-        {
-            sell_price = input_line.replace(".", "").parse().unwrap();
-            break;
-        } else {
-            print_error_line(terminal_io, "Invalid price entered, please retry!\n");
-        }
-    }
-    printline(terminal_io, "");
-    let box_count = loop {
-        utils::printline(terminal_io, "Enter how many boxes to add. Format: [0-9]+");
-        utils::printline(terminal_io, &format!("Modify or keep [0]:"));
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            break 0;
-        } else if Regex::new("^[0-9]+").expect("").is_match(&input_line) {
-            break input_line.parse().unwrap();
-        } else {
-            print_error_line(terminal_io, "Invalid stock entered, please retry!\n");
-        }
-    };
-    printline(terminal_io, "");
-    if box_count == 0 {
-        printline(terminal_io, "Added 0 boxes.");
-        return TimeoutResult::RESULT(());
-    }
-
-    match rv_api::buy_in_box(barcode, buy_price, sell_price, box_count, credentials).unwrap() {
-        ApiResult::Success => utils::printline(
-            terminal_io,
-            &format!(
-                "Added {} boxes. Total of {} items.",
-                box_count,
-                box_.items_per_box * box_count
-            ),
-        ),
-        ApiResult::Fail(msg) => print_error_line(terminal_io, &msg),
-    }
-
-    TimeoutResult::RESULT(())
-}
-
-fn buy_in_product(
-    barcode: &str,
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    let product;
-    match rv_api::get_product_info_admin(&credentials, barcode).unwrap() {
-        ApiResultValue::Success(suc) => product = suc,
-        ApiResultValue::Fail(msg) => {
-            utils::print_error_line(terminal_io, &msg);
-            return TimeoutResult::RESULT(());
-        }
-    };
-    utils::printline(terminal_io, &format!("Adding new products to stock."));
-    let mut buy_price = product.buy_price;
-    let mut buy_price_changed = false;
-    loop {
-        utils::printline(
-            terminal_io,
-            "Enter item buyprice. Format: [0-9]+\\.[0-9][0-9]",
-        );
-        utils::printline(terminal_io, "At least one number, followed by period, followed by two numbers. For example: '1.00', '0.01', '14.42'");
-        utils::printline(
-            terminal_io,
-            &format!("Modify or keep [{}]:", &utils::format_money(&buy_price)),
-        );
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            printline(terminal_io, "Nothing changed.");
-            break;
-        } else if Regex::new("^[0-9]+\\.[0-9][0-9]$")
-            .expect("")
-            .is_match(&input_line)
-        {
-            buy_price = input_line.replace(".", "").parse().unwrap();
-            buy_price_changed = true;
-            break;
-        } else {
-            print_error_line(terminal_io, "Invalid price entered, please retry!\n");
-        }
-    }
-    printline(terminal_io, "");
-
-    let mut sell_price = product.sell_price;
-    loop {
-        utils::printline(terminal_io, "\r\nEnter item sellprice.");
-        if buy_price_changed {
-            let margin = rv_api::get_margin(&credentials).unwrap() as f64;
-            let margin_pretty = format!("{}%", (margin * 100.0).ceil());
-            sell_price = (buy_price as f64 * (1.0 + margin)).ceil() as i32;
-            utils::printline(
-                terminal_io,
-                &format!(
-                    "Suggest {} calculated with the margin of {}",
-                    &utils::format_money(&sell_price),
-                    margin_pretty
-                ),
-            );
-            utils::printline(
-                terminal_io,
-                &format!("Modify or keep [{}]:", &utils::format_money(&sell_price)),
-            );
-        } else {
-            utils::printline(
-                terminal_io,
-                &format!("Modify or keep [{}]:", &utils::format_money(&sell_price)),
-            );
-        }
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            if buy_price_changed {
-                printline(terminal_io, "Using the suggested price.");
-            } else {
-                printline(terminal_io, "Nothing changed.");
-            }
-            break;
-        } else if Regex::new("^[0-9]+\\.[0-9][0-9]$")
-            .expect("")
-            .is_match(&input_line)
-        {
-            sell_price = input_line.replace(".", "").parse().unwrap();
-            break;
-        } else {
-            print_error_line(terminal_io, "Invalid price entered, please retry!\n");
-        }
-    }
-    printline(terminal_io, "");
-    let count = loop {
-        utils::printline(
-            terminal_io,
-            "How many products to add? Format: [0-9]+ or [0-9]+\\*[0-9]+",
-        );
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-
-        if input_line.len() == 0 {
-            break 0;
-        }
-
-        match utils::calculator_input(&input_line) {
-            Some(count) => {
-                break count;
-            }
-            None => {
-                print_error_line(terminal_io, "Invalid count entered, please retry!\n");
-            }
-        }
-    };
-    printline(terminal_io, "");
-    if count == 0 {
-        printline(terminal_io, "Added 0 products to stock.");
-        return TimeoutResult::RESULT(());
-    }
-
-    rv_api::buy_in_product(barcode, buy_price, sell_price, count, credentials);
-    utils::printline(terminal_io, &format!("Added {} products to stock.", count));
-    TimeoutResult::RESULT(())
-}
-
-fn return_purchase(
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    utils::print_title(terminal_io, "Return recent purchase");
-
-    utils::printline(terminal_io, "Enter product barcode:");
-    let barcode = match readline(terminal_io, INPUT_TIMEOUT_SHORT) {
-        TimeoutResult::RESULT(s) => {
-            if Regex::new("^[0-9]+$").unwrap().is_match(&s) {
-                s
-            } else {
-                print_error_line(terminal_io, "Invalid barcode!");
-
-                std::thread::sleep(std::time::Duration::from_millis(2000));
-                return TimeoutResult::RESULT(());
-            }
-        }
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-    };
-
-    match return_product(credentials, &barcode).unwrap() {
-        ApiResult::Success => {
-            let product = get_product_info(credentials, &barcode).unwrap();
-            printline(
-                terminal_io,
-                &format!("\nReturned product: {} successfully", product.name),
-            );
-        }
-        ApiResult::Fail(msg) => print_error_line(terminal_io, &format!("Return failed {msg}")),
-    }
-    TimeoutResult::RESULT(())
-}
-
-fn multibuy(
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    print_title(terminal_io, "Multibuy");
-
-    utils::printline(terminal_io, "Enter item barcode:");
-    let barcode = match readline(terminal_io, INPUT_TIMEOUT_LONG) {
-        TimeoutResult::RESULT(s) => {
-            if Regex::new("^[0-9]+$").unwrap().is_match(&s) {
-                s
-            } else {
-                print_error_line(terminal_io, "Invalid barcode!");
-                std::thread::sleep(std::time::Duration::from_millis(2000));
-                return TimeoutResult::RESULT(());
-            }
-        }
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-    };
-
-    utils::printline(terminal_io, "Enter item count to buy:");
-    let count: i32 = match readline(terminal_io, INPUT_TIMEOUT_LONG) {
-        TimeoutResult::RESULT(s) => {
-            if Regex::new("^[1-9][0-9]*$").unwrap().is_match(&s) {
-                s.parse().unwrap()
-            } else {
-                print_error_line(terminal_io, "Invalid count!");
-                std::thread::sleep(std::time::Duration::from_millis(2000));
-                return TimeoutResult::RESULT(());
-            }
-        }
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-    };
-    purchase_items(terminal_io, &barcode, count, credentials);
-    TimeoutResult::RESULT(())
-}
-
-fn purchase_items(
-    terminal_io: &mut TerminalIO,
-    barcode: &str,
-    count: i32,
-    credentials: &rv_api::AuthenticationResponse,
-) {
-    match rv_api::purchase_item(&credentials, &barcode, &count).unwrap() {
-        ApiResultPurchaseItem::Success => {
-            let product_info = rv_api::get_product_info(&credentials, &barcode).unwrap();
-            if product_info.barcode == "42615374" {
-                // Coffee purchase shill
-                utils::printline(terminal_io, &COFFEE_MSG);
-            }
-            utils::printline(
-                terminal_io,
-                &format!(
-                    "Bought {}x {} ({}EUR) Total ({}EUR)",
-                    count,
-                    product_info.name,
-                    utils::format_money(&product_info.price),
-                    utils::format_money(&(count * product_info.price))
-                ),
-            );
-        }
-        ApiResultPurchaseItem::Fail(x) => {
-            purchase_fail_bell();
-            let user_info = get_user_info(credentials).unwrap();
-            utils::set_small_font();
-            execute!(
-                terminal_io.writer,
-                PrintStyledContent(PURCHASE_FAILED_MSG1.to_string().green()),
-                PrintStyledContent(PURCHASE_FAILED_MSG2.to_string().red()),
-                Print("\r\n"),
-                Print(&format!("Dear {}, your purchase has", user_info.username)),
-                PrintStyledContent(" FAILED ".red()),
-                Print(&format!("with an error: {}\r\n", x.message))
-            )
-            .unwrap();
-            let wait_seconds = match x.error_type {
-                ApiResultPurchaseItemFailType::InsufficientFunds => 15,
-                _ => 5,
-            };
-            sleep(Duration::from_secs(wait_seconds));
-            Print(format!(
-                "You must wait {wait_seconds} seconds before you can proceed!\r\n"
-            ));
-            while terminal_io.recv.try_recv().is_ok() {
-                // Discard all input until channel is empty
-            }
-            utils::confirm_enter_to_continue(terminal_io);
-            utils::set_big_font();
-        }
-    }
-}
-
-fn search_products(
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    print_title(terminal_io, "Product search");
-    printline(terminal_io, "Enter name or barcode");
-    let query = match readline(terminal_io, INPUT_TIMEOUT_SHORT) {
-        TimeoutResult::RESULT(s) => s,
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-    };
-    let product_results = rv_api::search_products(credentials, &query).unwrap();
-    let user_info = get_user_info(credentials).unwrap();
-    let box_results = match user_info.is_admin() {
-        true => rv_api::search_boxes(credentials, &query).unwrap(),
-        false => Vec::new(),
-    };
-    if product_results.is_empty() && box_results.is_empty() {
-        utils::printline(
-            terminal_io,
-            &format!("No results found with query {}", &query),
-        );
-        return TimeoutResult::RESULT(());
-    }
-    printline(terminal_io, "\r\nResult products:");
-    let mut dupehack: Vec<String> = Vec::new();
-    for product in product_results {
-        dupehack.push(format!(
-            "{}, {} EUR, ID: {}, {} in stock.",
-            product.name,
-            utils::format_money(&product.price),
-            product.barcode,
-            product.stock
-        ));
-    }
-    for product in box_results.iter().map(|f| &f.product) {
-        dupehack.push(format!(
-            "{}, {} EUR, ID: {}, {} in stock.",
-            product.name,
-            utils::format_money(&product.sell_price),
-            product.barcode,
-            product.stock
-        ));
-    }
-    dupehack.sort();
-    dupehack.dedup();
-    for line in dupehack {
-        printline(terminal_io, &line);
-    }
-    if user_info.is_admin() {
-        printline(terminal_io, "\r\nResult boxes:");
-        for box_result in box_results {
-            utils::printline(
-                terminal_io,
-                &format!(
-                    "{} containing {}x of {} {}",
-                    box_result.box_barcode,
-                    box_result.items_per_box,
-                    box_result.product.barcode,
-                    box_result.product.name,
-                ),
-            );
-        }
-    }
-    TimeoutResult::RESULT(())
-}
-
-fn deposit(
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    print_title(terminal_io, "Deposit money");
-    utils::printline(
-        terminal_io,
-        "How much to deposit? Format: [0-9]+((\\.|,)[0-9][0-9])?",
-    );
-    utils::printline(terminal_io, "At least one number, optionally followed by a period or comma followed by two numbers. For example: '1', '0.10', '14,42'");
-    let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-        TimeoutResult::RESULT(s) => s,
-    };
-    if !Regex::new("^[0-9]+((\\.|,)[0-9][0-9])?$")
-        .unwrap()
-        .is_match(&input_line)
-    {
-        printline(terminal_io, "");
-        utils::print_error_line(terminal_io, "Invalid input. Deposit aborted!");
-        return TimeoutResult::RESULT(());
-    }
-    let amount: u32 = if input_line.contains(".") {
-        input_line.replace(".", "").parse().unwrap()
-    } else if input_line.contains(",") {
-        input_line.replace(",", "").parse().unwrap()
-    } else {
-        // No decimals specified, multiply by 100 to get cents
-        input_line.parse::<u32>().unwrap() * 100
-    };
-
-    if amount > 25000 {
-        printline(terminal_io, "");
-        utils::print_error_line(
-            terminal_io,
-            "You can deposit at most 250 EUR at once. Deposit aborted!",
-        );
-        return TimeoutResult::RESULT(());
-    }
-
-    let amount_formatted = utils::format_money(&(amount as i32));
-    execute!(
-        terminal_io.writer,
-        Print("\r\n"),
-        PrintStyledContent(
-            "PLEASE NOTE: WITHDRAWING MONEY IS NOT POSSIBLE."
-                .with(Color::Black)
-                .on(Color::White)
-        ),
-        Print("\r\n"),
-        Print(&format!(
-            "\
-            You can't transfer money to somebody else's account.\r\n\
-            Please confirm your deposit of {} euros.\r\n\
-            ",
-            amount_formatted
-        )),
-        Print("PLEASE TYPE '"),
-        PrintStyledContent(
-            amount_formatted
-                .to_string()
-                .with(Color::Black)
-                .on(Color::White)
-        ),
-        Print("' FOLLOWED BY <ENTER>: ")
-    )
-    .unwrap();
-
-    match utils::readline(terminal_io, INPUT_TIMEOUT_SHORT) {
-        TimeoutResult::TIMEOUT => {
-            utils::printline(terminal_io, "\r\nTimed out!");
-            std::thread::sleep(std::time::Duration::from_millis(2000));
-            return TimeoutResult::TIMEOUT;
-        }
-        TimeoutResult::RESULT(s) => {
-            if s.len() == 0 {
-                utils::printline(terminal_io, "\r\nDeposit aborted! Cancelled by user.");
-                return TimeoutResult::RESULT(());
-            } else if s.replace(",", ".") != amount_formatted {
-                utils::print_error_line(
-                    terminal_io,
-                    "\r\nDeposit aborted! Given amounts do not match.",
-                );
-                return TimeoutResult::RESULT(());
-            }
-        }
-    };
-
-    loop {
-        execute!(
-            terminal_io.writer,
-            Print("\r\n"),
-            Print("Did you deposit money as cash or via banktransfer?\r\n"),
-            Print("PLEASE TYPE EITHER '"),
-            PrintStyledContent("cash".with(Color::Black).on(Color::White)),
-            Print("' OR '"),
-            PrintStyledContent("bank".with(Color::Black).on(Color::White)),
-            Print("' FOLLOWED BY <ENTER>:\r\n"),
-        )
-        .unwrap();
-        match utils::readline(terminal_io, INPUT_TIMEOUT_SHORT) {
-            TimeoutResult::TIMEOUT => {
-                utils::printline(terminal_io, "\r\nTimed out!");
-                std::thread::sleep(std::time::Duration::from_millis(2000));
-                return TimeoutResult::TIMEOUT;
-            }
-            TimeoutResult::RESULT(s) => {
-                if s.len() == 0 {
-                    utils::printline(terminal_io, "\r\nDeposit aborted! Cancelled by user.");
-                    return TimeoutResult::RESULT(());
-                } else if s == "cash" {
-                    rv_api::deposit(&credentials, &amount, "cash").unwrap();
-                    utils::printline(terminal_io, "Remember to put cash in an envelope or send an email immediately to rahastonhoitaja@tko-aly.fi to explain a non-envelope deposit.");
-                    utils::printline(
-                        terminal_io,
-                        &format!("Current date: {}", Local::now().format("%d/%m/%Y")).to_string(),
-                    );
-
-                    utils::confirm_enter_to_continue(terminal_io);
-                    break;
-                } else if s == "bank" {
-                    rv_api::deposit(&credentials, &amount, "banktransfer").unwrap();
-                    break;
-                } else {
-                    print_error_line(terminal_io, "Invalid deposit type entered!");
-                }
-            }
-        };
-    }
-
-    utils::printline(
-        terminal_io,
-        &format!(
-            "\r\nDeposited {} EUR.",
-            utils::format_money(&(amount as i32))
-        ),
-    );
-    TimeoutResult::RESULT(())
-}
-
-fn change_username(
-    timeout: Duration,
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    print_title(terminal_io, "Change your username:");
-    execute!(terminal_io.writer, Print("New username: ")).unwrap();
-    let username = match utils::readline(terminal_io, timeout) {
-        TimeoutResult::TIMEOUT => {
-            return TimeoutResult::TIMEOUT;
-        }
-        TimeoutResult::RESULT(s) => s,
-    };
-
-    match rv_api::change_username(credentials, &username).unwrap() {
-        rv_api::ApiResult::Success => {
-            utils::printline(terminal_io, "Username successfully changed.");
-        }
-        rv_api::ApiResult::Fail(msg) => {
-            utils::print_error_line(terminal_io, &format!("Username change failed: {msg}"));
-        }
-    }
-
-    return TimeoutResult::RESULT(());
-}
-
-fn change_real_name(
-    timeout: Duration,
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    print_title(terminal_io, "Change your FULL name:");
-    execute!(terminal_io.writer, Print("Your FULL name: ")).unwrap();
-    let full_name = match utils::readline(terminal_io, timeout) {
-        TimeoutResult::TIMEOUT => {
-            return TimeoutResult::TIMEOUT;
-        }
-        TimeoutResult::RESULT(s) => s,
-    };
-
-    match rv_api::change_full_name(credentials, &full_name).unwrap() {
-        rv_api::ApiResult::Success => {
-            utils::printline(terminal_io, "Name successfully changed.");
-        }
-        rv_api::ApiResult::Fail(msg) => {
-            utils::print_error_line(terminal_io, &format!("Name change failed: {msg}"));
-        }
-    }
-    return TimeoutResult::RESULT(());
-}
-
-fn change_user_email(
-    timeout: Duration,
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    print_title(terminal_io, "Change Email");
-
-    execute!(terminal_io.writer, Print("Enter new email: ")).unwrap();
-    let email1;
-    match utils::readline(terminal_io, timeout) {
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-        TimeoutResult::RESULT(s) => email1 = s,
-    }
-
-    utils::printline(terminal_io, "");
-    execute!(terminal_io.writer, Print("Enter new email again: ")).unwrap();
-    let email2;
-    match utils::readline(terminal_io, timeout) {
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-        TimeoutResult::RESULT(s) => email2 = s,
-    }
-    utils::printline(terminal_io, "");
-
-    if email1.len() == 0 {
-        utils::printline(
-            terminal_io,
-            "Empty email is not allowed! Email not changed.",
-        );
-    } else if email1.split("@").count() != 2 {
-        utils::printline(terminal_io, "You did not provide valid email address");
-    } else if email1 == email2 {
-        match rv_api::change_email(credentials, &email1).unwrap() {
-            rv_api::ApiResult::Success => {
-                utils::printline(terminal_io, "Email successfully changed.");
-            }
-            rv_api::ApiResult::Fail(msg) => {
-                utils::print_error_line(terminal_io, &format!("Email change failed: {msg}"));
-            }
-        }
-    } else {
-        utils::printline(terminal_io, "Emails do not match! email not changed.");
-    }
-    utils::printline(terminal_io, "");
-    utils::confirm_enter_to_continue(terminal_io);
-    utils::printline(terminal_io, "");
-    TimeoutResult::RESULT(())
-}
-
-fn change_privacy(
-    timeout: Duration,
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    utils::print_title(terminal_io, "Privacy Settings");
-    printline(terminal_io, "Change the account's privacy level");
-    printline(terminal_io, "0 = No restrictions");
-    printline(
-        terminal_io,
-        "1 = Hide username from public (for example, leaderboards)",
-    );
-    printline(
-        terminal_io,
-        "2 = Hide all data from public (for example, list of recent purchases)",
-    );
-    printline(terminal_io, "<Enter> do not change");
-    loop {
-        match terminal_io.recv.recv_timeout(timeout) {
-            Ok(input::InputEvent::Terminal(Event::Key(ev))) => match ev.code {
-                KeyCode::Char(c) => match c {
-                    '0' => {
-                        rv_api::change_privacy_level(credentials, 0).unwrap();
-                        printline(terminal_io, "Changed privacy level to 0");
-                        return TimeoutResult::RESULT(());
-                    }
-                    '1' => {
-                        rv_api::change_privacy_level(credentials, 1).unwrap();
-                        printline(terminal_io, "Changed privacy level to 1");
-                        return TimeoutResult::RESULT(());
-                    }
-                    '2' => {
-                        rv_api::change_privacy_level(credentials, 2).unwrap();
-                        printline(terminal_io, "Changed privacy level to 2");
-                        return TimeoutResult::RESULT(());
-                    }
-                    _ => (),
-                },
-                KeyCode::Enter => return TimeoutResult::RESULT(()),
-                _ => (),
-            },
-            Err(RecvTimeoutError::Timeout) => return TimeoutResult::TIMEOUT,
-            _ => return TimeoutResult::RESULT(()),
-        }
-    }
-}
-
-fn change_user_password_user(
-    timeout: Duration,
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    print_title(terminal_io, "Change password");
-
-    execute!(terminal_io.writer, Print("Enter new password: ")).unwrap();
-    let password1;
-    match utils::readpasswd(terminal_io, timeout) {
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-        TimeoutResult::RESULT(s) => password1 = s,
-    }
-
-    utils::printline(terminal_io, "");
-    execute!(terminal_io.writer, Print("Enter new password again: ")).unwrap();
-    let password2;
-    match utils::readpasswd(terminal_io, timeout) {
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-        TimeoutResult::RESULT(s) => password2 = s,
-    }
-    utils::printline(terminal_io, "");
-
-    if password1.len() == 0 {
-        utils::printline(
-            terminal_io,
-            "Empty password is not allowed! Password not changed.",
-        );
-    } else if password1 == password2 {
-        match rv_api::change_password(credentials, &password1).unwrap() {
-            rv_api::ApiResult::Success => {
-                utils::printline(terminal_io, "Password successfully changed.");
-            }
-            rv_api::ApiResult::Fail(msg) => {
-                utils::print_error_line(terminal_io, &format!("Password change failed: {msg}"));
-            }
-        }
-    } else {
-        utils::printline(terminal_io, "Passwords do not match! Password not changed.");
-    }
-    utils::printline(terminal_io, "");
-    utils::confirm_enter_to_continue(terminal_io);
-    utils::printline(terminal_io, "");
-    TimeoutResult::RESULT(())
-}
-
-fn change_user_password_admin(
-    timeout: Duration,
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    print_title(terminal_io, "Change password (admin)");
-    execute!(terminal_io.writer, Print("Enter username: ")).unwrap();
-    let username;
-    match utils::readline(terminal_io, timeout) {
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-        TimeoutResult::RESULT(s) => username = s,
-    }
-
-    let user = match rv_api::get_user_info_by_username(credentials, &username).unwrap() {
-        ApiResultValue::Fail(msg) => {
-            print_error_line(terminal_io, &msg);
-            utils::printline(terminal_io, "");
-            return TimeoutResult::RESULT(());
-        }
-        ApiResultValue::Success(user) => user,
-    };
-
-    execute!(terminal_io.writer, Print("Enter new password: ")).unwrap();
-    let password1;
-    match utils::readpasswd(terminal_io, timeout) {
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-        TimeoutResult::RESULT(s) => password1 = s,
-    }
-
-    utils::printline(terminal_io, "");
-    execute!(terminal_io.writer, Print("Enter new password again: ")).unwrap();
-    let password2;
-    match utils::readpasswd(terminal_io, timeout) {
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-        TimeoutResult::RESULT(s) => password2 = s,
-    }
-    utils::printline(terminal_io, "");
-
-    if password1.len() == 0 {
-        utils::printline(
-            terminal_io,
-            "Empty password is not allowed! Password not changed.",
-        );
-    } else if password1 == password2 {
-        match rv_api::change_password_admin(credentials, user.user_id, &password1).unwrap() {
-            rv_api::ApiResult::Success => {
-                utils::printline(terminal_io, "Password successfully changed.");
-            }
-            rv_api::ApiResult::Fail(msg) => {
-                utils::print_error_line(terminal_io, &format!("Password change failed: {msg}"));
-            }
-        }
-    } else {
-        utils::printline(terminal_io, "Passwords do not match! Password not changed.");
-    }
-    utils::printline(terminal_io, "");
-    utils::confirm_enter_to_continue(terminal_io);
-    utils::printline(terminal_io, "");
-    TimeoutResult::RESULT(())
-}
-
-fn change_user_rfid(
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    print_title(terminal_io, "Set login RFID");
-    utils::printline(
-        terminal_io,
-        "Scan RFID to use for logging in. ENTER to cancel.",
-    );
-    loop {
-        match terminal_io.recv.recv_timeout(INPUT_TIMEOUT_SHORT) {
-            Err(RecvTimeoutError::Timeout) => return TimeoutResult::TIMEOUT,
-            Ok(input::InputEvent::Terminal(Event::Key(ev))) => match ev.code {
-                KeyCode::Enter => {
-                    utils::printline(terminal_io, "RFID change cancelled");
-                    return TimeoutResult::RESULT(());
-                }
-                _ => (),
-            },
-            Ok(input::InputEvent::Rfid(rfid)) => {
-                rv_api::change_rfid(credentials, &rfid).unwrap();
-                utils::printline(terminal_io, "RFID changed successfully");
-                return TimeoutResult::RESULT(());
-            }
-            _ => return TimeoutResult::RESULT(()),
-        }
-    }
-}
-
-fn change_item_properties(
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    print_title(terminal_io, "Change item properties");
-    utils::printline(terminal_io, "Enter barcode:");
-    let barcode = match readline(terminal_io, INPUT_TIMEOUT_LONG) {
-        TimeoutResult::RESULT(s) => {
-            if Regex::new("^[0-9]+$").unwrap().is_match(&s) {
-                s
-            } else {
-                print_error_line(terminal_io, "invalid barcode!");
-                std::thread::sleep(std::time::Duration::from_millis(2000));
-                return TimeoutResult::RESULT(());
-            }
-        }
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-    };
-
-    let product = match rv_api::get_product_info_admin(&credentials, &barcode).unwrap() {
-        ApiResultValue::Success(product) => Some(product),
-        ApiResultValue::Fail(msg) => {
-            utils::print_error_line(terminal_io, &msg);
-            None
-        }
-    };
-    if product.is_some() {
-        return change_product_properties(&barcode, terminal_io, credentials);
-    }
-    if let Some(b) = rv_api::get_box_info_admin(&barcode, credentials).unwrap() {
-        return change_box_properties(b.product.barcode, terminal_io, credentials);
-    }
-    utils::print_error_line(terminal_io, "No matching box or product found!");
-    TimeoutResult::RESULT(())
-}
-
-fn change_box_properties(
-    barcode: String,
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    let box_result = match rv_api::get_box_info_admin(&barcode, &credentials).unwrap() {
-        Some(b) => b,
-        None => {
-            print_error_line(terminal_io, &format!("No box found with {barcode}"));
-            return TimeoutResult::RESULT(());
-        }
-    };
-    let mut product_barcode = box_result.product.barcode;
-    printline(
-        terminal_io,
-        &format!("Current itembarcode: '{product_barcode}'"),
-    );
-    printline(terminal_io, &format!("Modify or keep [{product_barcode}]:"));
-    let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-        TimeoutResult::RESULT(s) => s,
-    };
-    if input_line.len() > 0 {
-        match get_product_info(credentials, &input_line) {
-            Some(_) => product_barcode = input_line,
-            None => match get_box_info_admin(&input_line, credentials).unwrap() {
-                Some(_) => {
-                    print_error_line(terminal_io, "Box with the given barcode already exists!");
-                    return TimeoutResult::RESULT(());
-                }
-                None => {
-                    if let TimeoutResult::TIMEOUT =
-                        new_product(&input_line, terminal_io, credentials)
-                    {
-                        return TimeoutResult::TIMEOUT;
-                    }
-                    if get_product_info(credentials, &input_line).is_none() {
-                        print_error_line(terminal_io, "Adding new item failed!");
-                        return TimeoutResult::RESULT(());
-                    }
-                    product_barcode = input_line;
-                }
-            },
-        }
-    } else {
-        printline(terminal_io, "Nothing changed.");
-    }
-
-    let mut items_per_box = box_result.items_per_box;
-    loop {
-        printline(
-            terminal_io,
-            &format!("Current items per box: '{items_per_box}'"),
-        );
-        printline(
-            terminal_io,
-            &format!("Modify or keep [{items_per_box}] Format: [0-9]+:"),
-        );
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            printline(terminal_io, "Nothing changed.");
-            break;
-        } else {
-            if Regex::new("^[0-9]+").expect("").is_match(&input_line) {
-                items_per_box = input_line.parse().unwrap();
-                break;
-            } else {
-                print_error_line(terminal_io, "Invalid number entered, please retry!\n");
-            }
-        }
-    }
-    match update_box(&barcode, items_per_box, &product_barcode, credentials) {
-        ApiResult::Success => printline(terminal_io, "Box modified successfully."),
-        ApiResult::Fail(msg) => {
-            print_error_line(terminal_io, &format!("Modifying box failed: {msg}"))
-        }
-    }
-    printline(terminal_io, "");
-
-    TimeoutResult::RESULT(())
-}
-
-fn change_product_properties(
-    barcode: &str,
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    let product = match rv_api::get_product_info_admin(&credentials, &barcode).unwrap() {
-        ApiResultValue::Success(product) => product,
-        ApiResultValue::Fail(msg) => {
-            utils::print_error_line(terminal_io, &msg);
-            return TimeoutResult::RESULT(());
-        }
-    };
-    let mut name = product.name;
-    utils::printline(terminal_io, &format!("Current description: '{name}'"));
-    utils::printline(terminal_io, &format!("Modify or keep [{name}]:"));
-
-    let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-        TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-        TimeoutResult::RESULT(s) => s,
-    };
-    if input_line.len() > 0 {
-        name = input_line.to_string();
-    } else {
-        printline(terminal_io, "Nothing changed.");
-    }
-
-    printline(terminal_io, "");
-
-    let mut buy_price = product.buy_price;
-    let mut buy_price_changed = false;
-    loop {
-        utils::printline(
-            terminal_io,
-            "Please enter item buyprice. Format: [0-9]+\\.[0-9][0-9]",
-        );
-        utils::printline(terminal_io, "At least one number, followed by period, followed by two numbers. For example: '1.00', '0.01', '14.42'");
-        utils::printline(
-            terminal_io,
-            &format!("Modify or keep [{}]:", &utils::format_money(&buy_price)),
-        );
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            printline(terminal_io, "Nothing changed.");
-            break;
-        } else if Regex::new("^[0-9]+\\.[0-9][0-9]$")
-            .expect("")
-            .is_match(&input_line)
-        {
-            buy_price = input_line.replace(".", "").parse().unwrap();
-            buy_price_changed = true;
-            break;
-        } else {
-            print_error_line(terminal_io, "Invalid price entered, please retry!\n");
-        }
-    }
-    printline(terminal_io, "");
-
-    let mut sell_price = product.sell_price;
-    loop {
-        utils::printline(terminal_io, "\r\nPlease enter item sellprice.");
-        if buy_price_changed {
-            let margin = rv_api::get_margin(&credentials).unwrap() as f64;
-            let margin_pretty = format!("{}%", (margin * 100.0).ceil());
-            sell_price = (buy_price as f64 * (1.0 + margin)).ceil() as i32;
-            utils::printline(
-                terminal_io,
-                &format!(
-                    "Suggest {} calculated with the margin of {}",
-                    &utils::format_money(&sell_price),
-                    margin_pretty
-                ),
-            );
-        }
-        utils::printline(
-            terminal_io,
-            &format!("Modify or keep [{}]:", &utils::format_money(&sell_price)),
-        );
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            if buy_price_changed {
-                printline(terminal_io, "Using the suggested price.");
-            } else {
-                printline(terminal_io, "Nothing changed.");
-            }
-            break;
-        } else if Regex::new("^[0-9]+\\.[0-9][0-9]$")
-            .expect("")
-            .is_match(&input_line)
-        {
-            sell_price = input_line.replace(".", "").parse().unwrap();
-            break;
-        } else {
-            print_error_line(terminal_io, "Invalid price entered, please retry!\n");
-        }
-    }
-    printline(terminal_io, "");
-    let mut stock = product.stock;
-    loop {
-        utils::printline(terminal_io, "Please enter item stock. Format: (+|-)?[0-9]+");
-        utils::printline(terminal_io, "No prefix or - means to set stock to negative or positive number given. + prefix means to increment stock e.g. +5 when stock is 2 results in stock of 7.");
-        utils::printline(terminal_io, &format!("Modify or keep [{stock}]:"));
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            printline(terminal_io, "Nothing changed.");
-            break;
-        } else if Regex::new("^(\\+|-)?[0-9]+")
-            .expect("")
-            .is_match(&input_line)
-        {
-            if input_line.chars().nth(0).unwrap() == '+' {
-                stock = stock + input_line.parse::<i32>().unwrap();
-            } else {
-                stock = input_line.parse().unwrap();
-            }
-            break;
-        } else {
-            print_error_line(terminal_io, "Invalid stock entered, please retry!\n");
-        }
-    }
-    printline(terminal_io, "");
-
-    let mut category = product.category;
-    loop {
-        utils::printline(terminal_io, "Please enter product category id.");
-        utils::printline(terminal_io, "Categories available:");
-        let categories = rv_api::get_categories(&credentials).unwrap();
-        for category in categories.iter() {
-            utils::printline(
-                terminal_io,
-                &format!("{}, id: {}", category.description, category.category_id),
-            );
-        }
-        utils::printline(
-            terminal_io,
-            &format!("Modify or keep [{}]:", category.description),
-        );
-        let input_line = match utils::readline(terminal_io, INPUT_TIMEOUT_LONG) {
-            TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-            TimeoutResult::RESULT(s) => s,
-        };
-        if input_line.len() == 0 {
-            printline(terminal_io, "Nothing changed.");
-            break;
-        } else if Regex::new("^[0-9]+").expect("").is_match(&input_line) {
-            let chosen: i32 = input_line.parse().unwrap();
-            match categories.iter().find(|c| c.category_id == chosen) {
-                Some(c) => {
-                    category = ProductCategory {
-                        description: c.description.clone(),
-                        category_id: c.category_id,
-                    };
-                }
-                None => {
-                    print_error_line(terminal_io, "Invalid category id entered, please retry!\n");
-                    continue;
-                }
-            }
-            break;
-        } else {
-            print_error_line(terminal_io, "Invalid category entered, please retry!\n");
-        }
-    }
-    rv_api::update_product(
-        &barcode,
-        &name,
-        category.category_id,
-        buy_price,
-        sell_price,
-        stock,
-        credentials,
-    )
-    .unwrap();
-    utils::printline(terminal_io, "Product updated.");
-    TimeoutResult::RESULT(())
-}
-
-fn management_mode_loop(
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    clear_terminal(terminal_io);
-    'main: loop {
-        let user_info = rv_api::get_user_info(&credentials).unwrap();
-
-        queue!(
-            terminal_io.writer,
-            cursor::MoveTo(0, terminal::size()?.1),
-            Print("=== management mode ===\r\n"),
-            PrintStyledContent("<barcode>".dark_green().bold()),
-            Print(" - IF FOUND update price and count ELSE add as a new item/box\r\n"),
-            PrintStyledContent("F".dark_green().bold()),
-            Print(" - list matching products\r\n"),
-            PrintStyledContent("I".dark_green().bold()),
-            Print(" - update all item/box properties\r\n"),
-            PrintStyledContent("U".dark_green().bold()),
-            Print(" - find an username\r\n"),
-            PrintStyledContent("P".dark_green().bold()),
-            Print(" - change password of an user\r\n"),
-            PrintStyledContent("E".dark_green().bold()),
-            Print(" - generate temppasword and send it to user\r\n"),
-            PrintStyledContent("<enter>".dark_green().bold()),
-            Print(" - exit management mode\r\n"),
-        )
-        .unwrap();
-
-        execute!(
-            terminal_io.writer,
-            Print(&format!(
-                "\nDear {}, your saldo is {} > ",
-                user_info.username,
-                utils::format_money(&user_info.money_balance)
-            ))
-        )
-        .unwrap();
-
-        let mut command = String::new();
-        loop {
-            match terminal_io.recv.recv_timeout(INPUT_TIMEOUT_LONG) {
-                Err(RecvTimeoutError::Timeout) => return TimeoutResult::TIMEOUT,
-                Ok(InputEvent::Terminal(Event::Key(ev))) => match ev.code {
-                    KeyCode::Char(c) => match c.to_ascii_lowercase() {
-                        'f' => {
-                            printline(terminal_io, "\n");
-                            match search_products(terminal_io, credentials) {
-                                TimeoutResult::TIMEOUT => break 'main,
-                                _ => (),
-                            }
-                            printline(terminal_io, "\n");
-                            break;
-                        }
-                        'i' => {
-                            printline(terminal_io, "");
-                            match change_item_properties(terminal_io, &credentials) {
-                                TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-                                TimeoutResult::RESULT(_) => (),
-                            }
-                            printline(terminal_io, "");
-                            continue 'main;
-                        }
-                        'u' => {
-                            printline(terminal_io, "\n");
-                            match find_user_admin(terminal_io, &credentials) {
-                                TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-                                TimeoutResult::RESULT(_) => (),
-                            }
-                            printline(terminal_io, "");
-                            break;
-                        }
-                        'p' => {
-                            printline(terminal_io, "\n");
-
-                            match change_user_password_admin(
-                                INPUT_TIMEOUT_LONG,
-                                terminal_io,
-                                &credentials,
-                            ) {
-                                TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-                                TimeoutResult::RESULT(_) => (),
-                            }
-                            printline(terminal_io, "");
-                            break;
-                        }
-                        'e' => {
-                            printline(terminal_io, "\n");
-
-                            match generate_temppassword_admin(
-                                INPUT_TIMEOUT_LONG,
-                                terminal_io,
-                                &credentials,
-                            ) {
-                                TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-                                TimeoutResult::RESULT(_) => (),
-                            }
-                            printline(terminal_io, "");
-                            break;
-                        }
-                        '0'..='9' => {
-                            terminal_io.writer.execute(Print(c)).unwrap();
-                            command.push(c);
-                        }
-                        _ => {}
-                    },
-                    KeyCode::Backspace => {
-                        if !command.is_empty() {
-                            execute!(
-                                terminal_io.writer,
-                                cursor::MoveLeft(1),
-                                Print(" "),
-                                cursor::MoveLeft(1)
-                            )
-                            .unwrap();
-                            command.pop();
-                        }
-                    }
-                    KeyCode::Enter => {
-                        command = command.trim().to_string();
-                        utils::printline(terminal_io, "\r\n");
-                        if command.is_empty() {
-                            clear_terminal(terminal_io);
-                            break 'main;
-                        } else if Regex::new("^[0-9]+$").expect("").is_match(&command) {
-                            let barcode = &command;
-                            if let Some(_) = rv_api::get_product_info(credentials, barcode) {
-                                match buy_in_product(barcode, terminal_io, credentials) {
-                                    TimeoutResult::RESULT(_) => (),
-                                    TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-                                }
-                                break;
-                            }
-                            if let Some(_) =
-                                rv_api::get_box_info_admin(barcode, credentials).unwrap()
-                            {
-                                match buy_in_box(barcode, terminal_io, credentials) {
-                                    TimeoutResult::RESULT(_) => (),
-                                    TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-                                }
-                                break;
-                            }
-                            print_error_line(
-                                terminal_io,
-                                &format!("No box or product found with barcode {barcode}"),
-                            );
-                            new_item(&command, terminal_io, credentials);
-                            break;
-                        } else {
-                            utils::print_error_line(
-                                terminal_io,
-                                &format!("unknown command: {}\r\n", &command),
-                            );
-                            break;
-                        }
-                    }
-                    KeyCode::F(5) => {
-                        break;
-                    }
-                    _ => (),
-                },
-                Ok(InputEvent::Rfid(_)) => {
-                    // Logout
-                    return TimeoutResult::RESULT(());
-                }
-                _ => (),
-            }
-        }
-    }
-    TimeoutResult::RESULT(())
-}
-
-fn settings_loop(
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) -> TimeoutResult<()> {
-    clear_terminal(terminal_io);
-    'main: loop {
-        let user_info = rv_api::get_user_info(&credentials).unwrap();
-
-        clear_terminal(terminal_io);
-
-        queue!(
-            terminal_io.writer,
-            cursor::MoveTo(0, terminal::size()?.1),
-            Print("Current values\r\n"),
-            PrintStyledContent("Name: ".dark_green().bold()),
-            Print(format!("{}\r\n", user_info.full_name)),
-            PrintStyledContent("Email: ".dark_green().bold()),
-            Print(format!("{}\r\n", user_info.email)),
-            PrintStyledContent("Privacy level: ".dark_green().bold()),
-            Print(format!("{}\r\n", user_info.privacy_level)),
-            Print("\r\n"),
-            Print("Available commands (press key to select):\r\n"),
-            PrintStyledContent("R".dark_green().bold()),
-            Print(" - manage your rfid\r\n"),
-            PrintStyledContent("P".dark_green().bold()),
-            Print(" - change your password\r\n"),
-            PrintStyledContent("E".dark_green().bold()),
-            Print(" - change your email\r\n"),
-            PrintStyledContent("N".dark_green().bold()),
-            Print(" - change your FULL name\r\n"),
-            PrintStyledContent("V".dark_green().bold()),
-            Print(" - change your privacy level\r\n"),
-        )
-        .unwrap();
-
-        if utils::is_barcode(&user_info.username) {
-            queue!(
-                terminal_io.writer,
-                PrintStyledContent("U".dark_green().bold()),
-                Print(" - change your username\r\n"),
-            )
-            .unwrap();
-        }
-
-        queue!(
-            terminal_io.writer,
-            PrintStyledContent("<enter>".dark_green().bold()),
-            Print(" - exit settings\r\n"),
-        )
-        .unwrap();
-
-        execute!(
-            terminal_io.writer,
-            Print(&format!(
-                "\nDear {}, your saldo is {} > ",
-                user_info.username,
-                utils::format_money(&user_info.money_balance)
-            ))
-        )
-        .unwrap();
-
-        let mut command = String::new();
-        loop {
-            match terminal_io.recv.recv_timeout(INPUT_TIMEOUT_LONG) {
-                Err(RecvTimeoutError::Timeout) => return TimeoutResult::TIMEOUT,
-                Ok(InputEvent::Terminal(Event::Key(ev))) => match ev.code {
-                    KeyCode::Char(c) => match c.to_ascii_lowercase() {
-                        'r' => {
-                            printline(terminal_io, "");
-                            match change_user_rfid(terminal_io, &credentials) {
-                                TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-                                TimeoutResult::RESULT(_) => (),
-                            }
-                            printline(terminal_io, "");
-                            break;
-                        }
-                        'v' => {
-                            printline(terminal_io, "");
-                            match change_privacy(INPUT_TIMEOUT_SHORT, terminal_io, &credentials) {
-                                TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-                                TimeoutResult::RESULT(_) => (),
-                            }
-                            printline(terminal_io, "");
-                            break;
-                        }
-                        'p' => {
-                            printline(terminal_io, "\n");
-                            match change_user_password_user(
-                                INPUT_TIMEOUT_LONG,
-                                terminal_io,
-                                &credentials,
-                            ) {
-                                TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-                                TimeoutResult::RESULT(_) => (),
-                            }
-                            printline(terminal_io, "");
-                            break;
-                        }
-                        'n' => {
-                            printline(terminal_io, "");
-                            match change_real_name(INPUT_TIMEOUT_LONG, terminal_io, &credentials) {
-                                TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-                                TimeoutResult::RESULT(_) => (),
-                            }
-                            printline(terminal_io, "");
-                            break;
-                        }
-                        'e' => {
-                            printline(terminal_io, "\n");
-                            match change_user_email(INPUT_TIMEOUT_LONG, terminal_io, &credentials) {
-                                TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-                                TimeoutResult::RESULT(_) => (),
-                            }
-                            printline(terminal_io, "");
-                            break;
-                        }
-                        'u' => {
-                            if utils::is_barcode(&user_info.username) {
-                                printline(terminal_io, "");
-                                match change_username(INPUT_TIMEOUT_LONG, terminal_io, &credentials)
-                                {
-                                    TimeoutResult::TIMEOUT => return TimeoutResult::TIMEOUT,
-                                    TimeoutResult::RESULT(_) => (),
-                                }
-                                printline(terminal_io, "");
-                            }
-                            break;
-                        }
-                        'q' => {
-                            // for jääräs
-                            clear_terminal(terminal_io);
-                            break 'main;
-                        }
-                        _ => {}
-                    },
-                    KeyCode::Backspace => {
-                        if !command.is_empty() {
-                            execute!(
-                                terminal_io.writer,
-                                cursor::MoveLeft(1),
-                                Print(" "),
-                                cursor::MoveLeft(1)
-                            )
-                            .unwrap();
-                            command.pop();
-                        }
-                    }
-                    KeyCode::Enter => {
-                        command = command.trim().to_string();
-                        utils::printline(terminal_io, "\r\n");
-                        if command.is_empty() {
-                            clear_terminal(terminal_io);
-                            break 'main;
-                        } else {
-                            utils::print_error_line(
-                                terminal_io,
-                                &format!("unknown command: {}\r\n", &command),
-                            );
-                            break;
-                        }
-                    }
-                    KeyCode::F(5) => {
-                        break;
-                    }
-                    _ => (),
-                },
-                _ => (),
-            }
-        }
-    }
-    TimeoutResult::RESULT(())
-}
-
-fn print_user_loop_instructions(
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) {
-    let user_info = rv_api::get_user_info(&credentials).unwrap();
-    queue!(
-        terminal_io.writer,
-        cursor::MoveTo(0, terminal::size()?.1),
-        Print("Available commands (press key to select):\r\n"),
-        PrintStyledContent("<barcode>".dark_green()),
-        Print(" - buy this item\r\n"),
-        PrintStyledContent("B".dark_green().bold()),
-        Print(" - buy item multiple times\r\n"),
-        PrintStyledContent("D".dark_green().bold()),
-        Print(" - deposit to your account\r\n"),
-        PrintStyledContent("F".dark_green().bold()),
-        Print(" - list matching products\r\n"),
-        PrintStyledContent("H".dark_green().bold()),
-        Print(" - show purchase history\r\n"),
-        PrintStyledContent("U".dark_green().bold()),
-        Print(" - undo a recent purchase\r\n"),
-        PrintStyledContent("S".dark_green().bold()),
-        Print(" - change settings\r\n"),
-        PrintStyledContent("C".dark_green().bold()),
-        Print(" - clear terminal\r\n"),
-        PrintStyledContent("<enter>".dark_green().bold()),
-        Print(" - log out\r\n"),
-    )
-    .unwrap();
-    utils::print_rv_logo(terminal_io);
-    if user_info.is_admin() {
-        queue!(
-            terminal_io.writer,
-            PrintStyledContent("M".dark_green().bold()),
-            Print(" - enter management mode\r\n"),
-        )
-        .unwrap();
-    }
-}
-
-fn print_user_loop_banner(
-    terminal_io: &mut TerminalIO,
-    credentials: &rv_api::AuthenticationResponse,
-) {
-    execute!(
-        terminal_io.writer,
-        terminal::Clear(terminal::ClearType::All)
-    )
-    .unwrap();
-    print_user_loop_instructions(terminal_io, credentials);
-    printline(terminal_io, "");
-}
-
-pub fn user_loop(terminal_io: &mut TerminalIO, credentials: &rv_api::AuthenticationResponse) {
-    print_user_loop_banner(terminal_io, credentials);
-
-    'main: loop {
-        let user_info = rv_api::get_user_info(&credentials).unwrap();
-        execute!(
-            terminal_io.writer,
-            Print(&format!(
-                "Dear {}, your saldo is {} > ",
-                user_info.username,
-                utils::format_money(&user_info.money_balance)
-            ))
-        )
-        .unwrap();
-
-        let mut command = String::new();
-        loop {
-            match terminal_io.recv.recv_timeout(INPUT_TIMEOUT_SHORT) {
-                Err(RecvTimeoutError::Timeout) => {
-                    utils::printline(terminal_io, "Timed out!");
-                    std::thread::sleep(std::time::Duration::from_millis(2000));
-                    break 'main;
-                }
-                Ok(input::InputEvent::Terminal(Event::Key(ev))) => match ev.code {
-                    KeyCode::Char(c) => match c.to_ascii_lowercase() {
-                        'b' => {
-                            printline(terminal_io, "\n");
-                            match multibuy(terminal_io, credentials) {
-                                TimeoutResult::TIMEOUT => break 'main,
-                                _ => (),
-                            }
-                            printline(terminal_io, "");
-                            break;
-                        }
-                        'd' => {
-                            printline(terminal_io, "\n");
-                            match deposit(terminal_io, &credentials) {
-                                TimeoutResult::TIMEOUT => break 'main,
-                                _ => (),
-                            }
-                            printline(terminal_io, "");
-                            break;
-                        }
-                        'f' => {
-                            printline(terminal_io, "\n");
-                            match search_products(terminal_io, credentials) {
-                                TimeoutResult::TIMEOUT => break 'main,
-                                _ => (),
-                            }
-                            printline(terminal_io, "");
-                            break;
-                        }
-                        'h' => {
-                            printline(terminal_io, "\n");
-                            print_title(terminal_io, "Recent purchases");
-                            let mut events = rv_api::purchase_history(credentials);
-                            events.sort_by(|a, b| b.time.cmp(&a.time));
-                            events.iter().take(10).rev().for_each(|event| {
-                                printline(
-                                    terminal_io,
-                                    &format!(
-                                        "{}{}{} {}€",
-                                        DateTime::parse_from_rfc3339(&event.time)
-                                            .unwrap()
-                                            .with_timezone(&chrono_tz::Europe::Helsinki)
-                                            .format("%d/%m/%Y %H:%M"),
-                                        if event.returned {
-                                            " bought [returned] "
-                                        } else {
-                                            " bought "
-                                        },
-                                        event.product.name,
-                                        utils::format_money(&event.price)
-                                    ),
-                                )
-                            });
-                            printline(terminal_io, "");
-                            break;
-                        }
-                        'm' => {
-                            if user_info.is_admin() {
-                                printline(terminal_io, "\n");
-                                match management_mode_loop(terminal_io, credentials) {
-                                    TimeoutResult::TIMEOUT => break 'main,
-                                    _ => (),
-                                }
-                                print_user_loop_instructions(terminal_io, credentials);
-                                break;
-                            }
-                        }
-                        's' => {
-                            printline(terminal_io, "\n");
-                            match settings_loop(terminal_io, credentials) {
-                                TimeoutResult::TIMEOUT => break 'main,
-                                _ => (),
-                            }
-                            print_user_loop_instructions(terminal_io, credentials);
-                            break;
-                        }
-                        'u' => {
-                            printline(terminal_io, "\n");
-                            match return_purchase(terminal_io, credentials) {
-                                TimeoutResult::TIMEOUT => break 'main,
-                                _ => (),
-                            }
-                            printline(terminal_io, "");
-                            break;
-                        }
-                        'q' => {
-                            // Legacy behavior wanted by some old users, need not to show in the list of commands
-                            break 'main; // Logout
-                        }
-                        'c' => {
-                            // Clear current terminal view
-                            // Useful after registering, if you want to see the list of commands
-                            // after logging in
-                            break print_user_loop_banner(terminal_io, credentials);
-                        }
-                        '0'..='9' => {
-                            terminal_io.writer.execute(Print(c)).unwrap();
-                            command.push(c);
-                        }
-                        _ => (),
-                    },
-                    KeyCode::Backspace => {
-                        if !command.is_empty() {
-                            execute!(
-                                terminal_io.writer,
-                                cursor::MoveLeft(1),
-                                Print(" "),
-                                cursor::MoveLeft(1)
-                            )
-                            .unwrap();
-                            command.pop();
-                        }
-                    }
-                    KeyCode::Enter => {
-                        command = command.trim().to_string();
-                        utils::printline(terminal_io, "\r\n");
-                        if command.is_empty() {
-                            break 'main; // Logout
-                        } else if command == "exit" {
-                            disable_raw_mode().unwrap();
-                            exit(0);
-                        } else if Regex::new("^[0-9]+$").expect("").is_match(&command) {
-                            purchase_items(terminal_io, &command, 1, credentials);
-                            printline(terminal_io, "");
-                            break;
-                        } else {
-                            utils::print_error_line(
-                                terminal_io,
-                                &format!("unknown command: {}\r\n", &command),
-                            );
-                            break;
-                        }
-                    }
-                    KeyCode::F(5) => {
-                        break;
-                    }
-                    _ => (),
-                },
-                Ok(InputEvent::Rfid(_)) => {
-                    // Logout
-                    return;
-                }
-                _ => (),
-            };
-        }
-    }
-}
